\name{mean_max_temps}
\docType{data}
\alias{mean_max_temps}
\title{Mean Maximum Tempeartures in Europe}
\description{
	This data has a response Y that is the mean of the annual
maximum temperatues from 1990-2010. Includes spatial coordinates
and elevation. Has 942 observations containing a large cluster
of observations in Germany.
}
\usage{mean_max_temps}
\format{
	Data frame of 942 observations with temperature (temp),
country (country), elevation (elev), latitude (lat), and longitude (long).
}
\source{Unknown}
\keyword{datasets}
