\name{plot.spacious}
\alias{plot.spacious}
\title{Diagnostic plot of spacious fit}
\description{
Plots covariance parameter estimates and log-likelihood for each Fisher
scoring iteration to help diagnose convergence issues.
}
\usage{
\method{plot}{spacious}(x, ...)
}
\arguments{
	\item{x}{model fit with \code{\link{spacious}}.}
	\item{...}{further arguments passed to or from other methods.}
}
\seealso{
	\code{\link{spacious}}
}
