\name{spacious-package}
\alias{spacious-package}
\docType{package}
\title{
Block composite spatial analysis.
}
\description{
Fits block composite geostatistical spatial models.
}
\details{
\tabular{ll}{
Package: \tab spacious\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2013-01-14\cr
License: \tab GPL\cr
}
spacious() is important.
}
\author{
Ryan J. Parker, Jarad B. Niemi, and Brian J. Reich

Maintainer: Ryan J. Parker <rjparker@ncsu.edu>
}
\references{
Jo Eidsvik, Benjamin A. Shaby, Brian J. Reich, Matthew Wheeler, and Jarad Niemi.
``Estimation and prediction in spatial models with block composite likelihoods using
parallel computing.'' Journal of Computational and Graphical Statistics (accepted)
}
\keyword{ spatial }
\seealso{
	\code{\link{spacious}}, \code{\link{mean_max_temps}}
}
