\name{spacious.fit}
\alias{spacious.fit}
\title{Fit block composite model}
\description{
	Fits a block composite model.
}
\usage{
	spacious.fit(y, X, D, B, neighbors, cov, n, p, R, theta)
}
\arguments{
	\item{y}{response}
}
\seealso{
	\code{\link{spacious}}
}
\examples{
# example here
}
\keyword{spacious.fit}
\keyword{internal}
