\name{summary.spacious}
\alias{summary.spacious}
\title{Summarize Spacious Objects}
\description{
	Summarize a model fit with spacious
}
\usage{
	\method{summary}{spacious}(object, \dots)
}
\arguments{
	\item{object}{a model fit with spacious.}
	\item{\dots}{other arguments}
}
\seealso{
	\code{\link{spacious}}
}
\examples{
# example here
}
\keyword{summary}
