\name{spacious-package}
\alias{spacious-package}
\docType{package}
\title{
Block composite spatial analysis.
}
\description{
Fits block composite geostatistical spatial models.
}
\details{
\tabular{ll}{
Package: \tab spacious\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2012-07-30\cr
License: \tab GPL\cr
}
spacious() is important.
}
\author{
Ryan J. Parker and Jarad Niemi

Maintainer: Ryan J. Parker <rjparker@ncsu.edu>
}
\references{
Surely we can cite some stuff.
}
\keyword{ spatial }
\seealso{
Something
}
\examples{
# Here are some examples
}
